added testMode, made forEach into for loops
testMode to make having the bot running locally & remotely less of an issue

+++ listeners/whitelistRemoveOnLeave.js
@@ -13,21 +13,23 @@ class WhitelistRemoveOnLeave extends Listener {
 	}
 
 	async exec(member) {
-		let ch_whitelist, ch_console;
-		member.guild.channels.forEach(async function(channel) {
-			if (channel.name == "whitelist") {
-				ch_whitelist = channel;
-			} else if (channel.name == "server-console") {
-				ch_console = channel;
+		if (!this.client.testMode) {
+			let ch_whitelist, ch_console;
+			for (let channel of member.guild.channels) {
+				if (channel[1].name == "whitelist") {
+					ch_whitelist = channel[1];
+				} else if (channel[1].name == "server-console") {
+					ch_console = channel[1];
+				}
 			}
-		});
 
-		if (ch_whitelist && ch_console) {
-			let messages = await ch_whitelist.fetchMessages(100);
-			for (let m of messages) {
-				if (m[1].author.id == member.id) {
-					await ch_console.send(`whitelist remove ${m[1].content}`)
-					await m[1].delete()
+			if (ch_whitelist && ch_console) {
+				let messages = await ch_whitelist.fetchMessages(100);
+				for (let m of messages) {
+					if (m[1].author.id == member.id) {
+						await ch_console.send(`whitelist remove ${m[1].content}`)
+						await m[1].delete()
+					}
 				}
 			}
 		}
